requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pydantic>=2.9
python-dotenv>=1.0
requests>=2.32
google-genai>=0.3
httpx>=0.27
